+++ external/vulkancts/modules/vulkan/video/vktVideoDecodeTests.cpp
@@ -526,7 +526,7 @@ public:
 		DE_ASSERT(m_decoder->m_cachedDecodeParams.size() == framesToDecode);
 	}
 
-	int getBufferedDisplayCount() const { return m_decoder->GetVideoFrameBuffer()->GetDisplayedFrameCount(); }
+	int getBufferedDisplayCount() const { return (int)m_decoder->GetVideoFrameBuffer()->GetDisplayedFrameCount(); }
 private:
 	DeviceContext* m_devctx;
 	ESEDemuxer m_demuxer;
@@ -934,7 +934,7 @@ tcu::TestStatus InterleavingDecodeTestInstance::iterate(void)
 	{
 		if (!res.incorrectFrames.empty())
 			allTestsPassed = false;
-		totalFramesCheck += (res.correctFrames.size() + res.incorrectFrames.size());
+		totalFramesCheck += (int)(res.correctFrames.size() + res.incorrectFrames.size());
 	}
 	DE_ASSERT(totalFramesCheck == totalFrames);
 	DE_UNREF(totalFramesCheck);
